requirements.txt
napari-plugin-engine>=0.1.4
numpy
pyyaml
scikit-image
tensorstore
zarr
